requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
jinja2==3.1.4
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
apscheduler==3.10.4
google-genai==1.0.0
aiogram==3.13.0
httpx==0.27.0
